requirements.txt
fastapi
uvicorn
requests
python-multipart
pillow
jinja2
pydantic
